refactor(kwwk): move stored-provider registration into KWWKAI
Lift the OAuth-store-backed provider layer out of the CLI target so
library consumers (import KWWKAI) can turn a synced oauth.json store
into live providers the same way the CLI does:

- ProviderDirectory.swift moves to KWWKAI as-is, everything public.
- New KWWKAI/StoredProviders.swift holds ResolvedAuth, ProviderSlot,
  SessionAuthResolvers, AuthResolveError, registerAllStored (was
  private in the CLI) and registerStored with all per-provider
  registration functions.
- Library code no longer prints: the two stderr notices become a
  'notice' callback (default no-op); the CLI passes cliAuthNotice to
  keep the exact same stderr output.
- registerAllStored now reads store.all() itself instead of taking a
  redundant 'all' parameter.
- CLI keeps loggedOutModel, resolveAgentAuth, registerStoredProviderLive,
  the env-key fallback path, and the TUI-side SessionProviders class.

No behavior change; 1315 tests pass unchanged.
